+++ symbol_by_market_phase.py
@@ -1,6 +1,6 @@
 import config
 import pandas as pd
-from datetime import datetime
+from datetime import datetime, timedelta
 import numpy as np
 import sys
 import timeit
@@ -8,7 +8,7 @@ import add_symbol
 import telegram
 import logging
 import database
-from exchange import client, get_exchange_info
+from exchange import client, get_exchange_info, get_price_close_by_symbol_and_date
 
 # calculate program run time
 start = timeit.default_timer() 
@@ -114,6 +114,84 @@ def get_data(symbol):
         frame = pd.DataFrame()
         return frame 
 
+def create_balance_snapshot():
+    msg = "Creating balance snapshot"
+    print(msg)
+    telegram.send_telegram_message(telegram.telegram_token_market_phases, "", msg)
+
+    last_date = database.get_last_date_from_balances(database.conn)
+    if last_date == '0':
+         today = datetime.now()
+         start_date = today - timedelta(days=30)
+    else:
+         start_date = datetime.strptime(last_date, '%Y-%m-%d')        
+
+    snapshots = client.get_account_snapshot(type="SPOT",
+                                            startTime=int(start_date.timestamp()*1000), 
+                                            # endTime=int(end_date.timestamp()*1000)
+                                            limit=30 #max                                        
+                                            )
+
+    code = snapshots['code']
+    msg = snapshots['msg']
+
+    # get list of available symbols. 
+    # This is usefull to avoid getting price from symbol that do not trade against stable
+    exchange_info = get_exchange_info()
+    symbols = set()
+    trade_against = "BUSD"
+    for s in exchange_info['symbols']:
+        if (s['symbol'].endswith(trade_against)
+            and s['status'] == 'TRADING'):
+                symbols.add(s['symbol']) 
+
+    # Create a Pandas DataFrame to store the daily balances for each asset
+    df_balance = pd.DataFrame()
+
+    # Iterate through the snapshots and get the daily balance for each asset
+    for snapshot in snapshots['snapshotVos']:
+        if snapshot['type'] == 'spot' and snapshot['data'] is not None:
+            snapshot_date = datetime.fromtimestamp(snapshot['updateTime']/1000).date()
+            totalAssetOfBtc = snapshot['data']['totalAssetOfBtc']
+            for balance in snapshot['data']['balances']:
+                asset = balance['asset']
+                daily_balance = float(balance['free'])
+
+                print(f"{snapshot_date}-{asset}")
+                
+                # ignore if balance = 0
+                if daily_balance == 0.0:
+                    continue
+
+                symbol_with_trade_against = asset+trade_against
+
+                if asset in [trade_against]:
+                     balance_usd = daily_balance
+                elif symbol_with_trade_against not in symbols:
+                     print(f"{asset} not in available symbols")
+                     balance_usd = 0
+                else:
+                    # convert snapshot_date from date to datetime
+                    date = datetime.combine(snapshot_date, datetime.min.time())
+                    unit_price = get_price_close_by_symbol_and_date(symbol_with_trade_against, date)
+                    balance_usd = unit_price * daily_balance
+
+                df_new = pd.DataFrame({
+                    'Date': [snapshot_date],
+                    'Asset': [asset],
+                    'Balance': [daily_balance],
+                    'Balance_USD': [balance_usd],
+                    'Total_Balance_Of_BTC': [totalAssetOfBtc]
+                    })
+                # add to total
+                df_balance = pd.concat([df_balance, df_new], ignore_index=True)
+
+    # Print the daily balances for each asset
+    # print(df_balance)
+
+    # add data to table Balance
+    database.add_balances(database.conn, df_balance)
+
 # empty dataframe
 dfResult = pd.DataFrame()
 
@@ -190,9 +268,7 @@ telegram.send_telegram_message(telegram.telegram_token_market_phases, "", df_top
 
 # create file to import to TradingView with the list of top performers and symbols in position 
 df_tv_list = database.get_distinct_symbol_by_market_phase_and_positions(database.conn)
-
 df_top = df_tv_list
-
 df_tv_list['symbol'] = "BINANCE:"+df_tv_list['symbol']
 # Write DataFrame to CSV file
 filename = "Top_performers_"+trade_against+".txt" 
@@ -202,8 +278,10 @@ telegram.send_telegram_message(telegram.telegram_token_market_phases, "", msg)
 telegram.send_telegram_file(telegram.telegram_token_market_phases, filename)
 #---------------------------------------------
 
-if not df_top.empty:
+# save each symbol balance to database to be used on dashboard charts
+create_balance_snapshot()
 
+if not df_top.empty:
     # remove coins from position files that are not top performers in accumulation or bullish phase
     database.delete_positions_not_top_rank(database.conn)